Рефакторинг/requirements.txt
fastapi==0.116.1
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
sqlalchemy==2.0.47
asyncpg==0.31.0
loguru==0.7.3
slowapi==0.1.9
bcrypt==5.0.0
aiosmtplib==5.1.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
websockets==16.0
jinja2==3.1.6
limits==5.8.0
pyasn1==0.6.2
rsa==4.9.1
ecdsa==0.19.1